feat(default labels): Add region to label if different from city
This adds the region to the default labels, but only if the _region
name_ is different from the _city name_ (defined as locality or
localadmin name).

The intent is to handle major world cities like Berlin, Sao Paulo,
Paris, etc that are contained within an administrative region of the
same name, and are so well known that they do not require any additional
specifiers.

Differences such as capitalization and accents are ignored for
comparison purposes, since they often exist in real data.

In the more common case where the region and city names are different,
the region abbreviation is preferred, with the region name being
returned only if the abbreviaton is not available.