feat(Singapore labels): support likely futre data format
As being discussed in
https://github.com/whosonfirst-data/whosonfirst-data/issues/1092, Who's
on First will likely include Singapore districts in either the
`macrohood` or `neighbourhood` layer eventually.

Once available, we definitely want these in the labels, so the label
schema is adjusted so that only the country will appear for now, but the
district name will be included once available.