feat: Add the concept of a 'label builder'
This allows us to control higher level aspects of the label, such as
whether or not administrative data comes first.

Most parts of the world would write an address with the
housenumber/street first, and admin info second, but not _all_ places do
this.

With a label builder we can support these places